contacts sesessions working for all routes

+++ models/contactsTable.js
@@ -4,33 +4,24 @@ const addContact = (contactName, companyName, email, phoneNumber, notes, userId)
   return db.query('INSERT INTO contacts (contact_name, company_name, email, phone_number, notes, user_id) VALUES ($1, $2, $3, $4, $5, $6)', [contactName, companyName, email, phoneNumber, notes, userId])
     .then((result) => result)
 }
-// const addContact = (contactName, companyName, email, phoneNumber, notes) => {
-//   return db.query('INSERT INTO contacts (contact_name, company_name, email, phone_number, notes) VALUES ($1, $2, $3, $4, $5)', [contactName, companyName, email, phoneNumber, notes])
-//     .then((result) => result)
-// }
 
 const getAllContacts = (userId) => {
   return db.query('SELECT * FROM contacts WHERE user_id =$1 ORDER BY contact_name ASC', [userId])
     .then((result) => result.rows)
 }
 
-// const getAllContacts = () => {
-//   return db.query('SELECT * FROM contacts ORDER BY contact_name ASC')
-//     .then((result) => result.rows)
-// }
-
-const getContactById = (id) => {
-  return db.query('SELECT * FROM contacts WHERE id =$1', [id])
+const getContactById = (id, userId) => {
+  return db.query('SELECT * FROM contacts WHERE id =$1 AND user_id = $2', [id, userId])
     .then((result) => result.rows[0])
 }
 
-const deleteContactById = (id) => {
-  return db.query('DELETE FROM contacts WHERE id =$1', [id])
+const deleteContactById = (id, userId) => {
+  return db.query('DELETE FROM contacts WHERE id =$1 AND user_id = $2', [id, userId])
     .then((result) => result.rows[0])
 }
 
-const updateContact = (id, contactName, companyName, email, phoneNumber, notes) => {
-  return db.query('UPDATE contacts SET contact_name = $2, company_name = $3, email = $4, phone_number = $5, notes = $6 WHERE id=$1', [id, contactName, companyName, email, phoneNumber, notes])
+const updateContact = (id, userId, contactName, companyName, email, phoneNumber, notes, ) => {
+  return db.query('UPDATE contacts SET contact_name = $3, company_name = $4, email = $5, phone_number = $6, notes = $7 WHERE id=$1 AND user_id = $2', [id, userId, contactName, companyName, email, phoneNumber, notes])
     .then((result) => result.rows[0])
 }
